requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask~=2.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv~=1.0.1
Werkzeug==3.1.3

SQLAlchemy~=2.0.44
mysql-connector-python
requests~=2.32.0
matplotlib~=3.10.3
gunicorn==23.0.0